Fix payments.amount type

+++ challenge/forms.py
@@ -1,11 +1,10 @@
-from wtforms import Form
-from wtforms import StringField, ValidationError, IntegerField
+from wtforms import Form, FloatField, StringField, ValidationError
 from wtforms.validators import Optional
 
 
 class GetPatientsForm(Form):
-    payment_min = IntegerField('payment_min', validators=[Optional()])
-    payment_max = IntegerField('payment_max', validators=[Optional()])
+    payment_min = FloatField('payment_min', validators=[Optional()])
+    payment_max = FloatField('payment_max', validators=[Optional()])
 
     def validate_payment_max(self, value):
         if not self.payment_min.data or not value.data:
